refactor: move table functions to table.rs thanks to @panarch
